backend/requirements.txt
flask==2.1.2
pyjwt==2.4.0
datetime
uuid
flask-mongoengine==1.0.0
Flask-RESTful==0.3.9
Flask-JWT-Extended==4.4.1
Flask-Bcrypt==1.0.1
